Add Dynamic Home Page Link to Menu
Each of our menu include files had hard-coded URLs to the home page, which makes navigating in a development environment a pain. I updated the GIOS and CAP menu files to use values from the context file for these URLs. That way, each 'Data' menu entry can point to my local home page, and we can update the context files on deployment to point them to the correct URLs when we go live.

* Updated both CAP and GIOS header/menu files to use a dynamic value for the URL of the 'Data' menu item

* Updated the context-reader.jsp file (our hub for reading context file values) to bring in the new homePageUrl value

* Modified the menu include on all user-facing pages to pass that value into the included menu file