Add New PNG to Replace SVG
We had an SVG on the home page(s) of the portal, and I believe some Tomcat and/or server rule was not allowing us to upload SVG images because they are executable files. Forgot about this, and when we moved to the new code base I ended up with missing images on the home page.

I've made a PNG file at a fixed size to replace the SVG, and updated the home page content area includes to reference that new PNG.