Update Background 'Hero' Images for Both Sites
Fixed an issue where both sites were using the same CSS file, causing both of the hero images to be the same - and not allowing us to use different images for the CAP and GIOS sites.

Although not optimal, I just ended up using inline style declarations within each of the 'content' includes. This way, each hero image can be declared right within the content file that is displaying it. Removed the CSS rule for this <div> from the main stylesheet.

If time ever permits, we could refactor this so that each site loads it's own CSS file (in conjunction with a global one for shared styles?).